Implement GraphQL user creation, `UserAuthority.Update`, and make controller fully authority reliant
